+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 
+# v0.3
+ * Add command line option flags
+
+# v0.2
+ * Add spreadsheet generation
+
 # v0.1
  * Write total working hours to stdout